requirements.txt
fastapi>=0.116.0,<1.0.0
uvicorn[standard]>=0.35.0,<1.0.0
sqlalchemy[asyncio]>=2.0.41,<3.0.0
asyncpg>=0.30.0,<1.0.0
alembic>=1.16.0,<2.0.0
psycopg[binary]>=3.2.0,<4.0.0
pydantic-settings>=2.10.0,<3.0.0
python-jose[cryptography]>=3.5.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
email-validator>=2.2.0,<3.0.0
google-auth>=2.40.0,<3.0.0
requests>=2.32.0,<3.0.0
httpx>=0.28.0,<1.0.0
python-multipart>=0.0.20,<1.0.0
